@@ -1,27 +1,10 @@
-import { directories, getAllDirectories } from "@/services/directories";
-import {
-  batch,
-  createContext,
-  onMount,
-  useContext,
-  type Component,
-  type JSX,
-} from "solid-js";
-import {
-  createStore,
-  type StoreSetter,
-  type SetStoreFunction,
-} from "solid-js/store";
+import { ISearchState } from "@/stores/store";
+import { createContext, useContext, type Component, type JSX } from "solid-js";
+import { type SetStoreFunction } from "solid-js/store";
 
 interface Collapsed {
   rootDirectory: string;
   collapsed: string[];
-}
-
-export interface SearchState {
-  collapsed: Collapsed[];
-  search: string;
-  pathSelected: string;
 }
 
 interface SearchActions {
@@ -30,47 +13,24 @@ interface SearchActions {
   setPathSelected: SetStoreFunction<string>;
 }
 
-type SearchStore = [SearchState, SearchActions];
-
-const initialSearchStore: SearchState = {
-  collapsed: directories.map((dir) => ({
-    rootDirectory: dir.path,
-    collapsed: [],
-  })),
-  search: "",
-  pathSelected: "",
-};
-
+type SearchStore = [ISearchState, SearchActions];
 const SearchContext = createContext<SearchStore>();
 
 export const SearchProvider: Component<{ children: JSX.Element }> = (props) => {
-  onMount(() => {
-    void (async () => {
-      await getAllDirectories();
-    })();
-  });
-
-  const [store, setStore] = createStore(initialSearchStore);
-
-  const searchStore = [
-    store,
-    {
-      setCollapse: (...args: Collapsed[]) => {
-        batch(() => {
-          setStore("collapsed", ...(args as [any]));
-        });
-      },
-      setSearch: (value: StoreSetter<string, ["search"]>) => {
-        setStore("search", value);
-      },
-      setPathSelected: (path: StoreSetter<string, ["pathSelected"]>) => {
-        setStore("pathSelected", path);
-      },
-    },
-  ] as SearchStore;
-
+  console.log("SearchProvider");
   return (
-    <SearchContext.Provider value={searchStore}>
+    <SearchContext.Provider
+      value={
+        [
+          props.item.data,
+          {
+            setCollapse: props.item.setCollapse,
+            setSearch: props.item.setSearch,
+            setPathSelected: props.item.setPathSelected,
+          },
+        ] as SearchStore
+      }
+    >
       {props.children}
     </SearchContext.Provider>
   );
